* s_system is no longer in userinfos * fixed warnings for FS_WriteFile (int => size_t) * changed skills from uint8_t to int16_t and use MSG_ReadShort and MSG_WriteShort to submit them * re.Init (R_Init) will return true on success and false on failure - not -1 * changed a strncpy to Q_strncpyz in sv_main.c * some smaller doxygen comments * some new debug messages to determine where the map/server start problem is

+++ src/client/ref.h
@@ -344,7 +344,7 @@ typedef struct {
 	/* or a discrete file from anywhere in the quake search path */
 	/* a -1 return means the file does not exist */
 	/* NULL can be passed for buf to just determine existance */
-	int (*FS_WriteFile) (const void *buffer, int len, const char *filename);
+	int (*FS_WriteFile) (const void *buffer, size_t len, const char *filename);
 	int (*FS_LoadFile) (char *name, void **buf);
 	void (*FS_FreeFile) (void *buf);
 	int (*FS_CheckFile) (const char *name);
